refactor(vitest): split main and worker instrumentation
